Refactor extensions
